requirements.txt
streamlit
asyncio
asyncpg
pgvector
openai
python-dotenv
psycopg2
httpx
loguru
